[LRA-3]Get Product By id: - get method was placed in services/http.ts file
