+++ src/services/productAPI.ts
@@ -1,18 +1,8 @@
 import api from '../configs/api';
-
-const handleError = (res: any) => {
-  if (!res.ok) {
-    throw Error(res.statusText);
-  }
-  return res.json();
-}
+import { get } from './http';
 
 export const productAPI = {
-  getProductByID: (product_id: string) => (
-    fetch(`${api.host}/products/${product_id}`)
-      .then(res => handleError(res))
-      .then(res => res.data)
-  )
+  getProductByID: (product_id: string) => get(`${api.host}/products/${product_id}`),
 }
 
 export default productAPI;
